added sweet alert for deletion

+++ public/modules/cameras/controllers/camera_list.js
@@ -1,7 +1,7 @@
 'use strict';
 
 angular.module('app')
-    .controller('CameraListCtrl', function ($scope, cameraService, ngTableParams, $modal, $log, psResponsive,$window) {
+    .controller('CameraListCtrl', function ($scope, cameraService, ngTableParams, $modal, $log, psResponsive,$window, SweetAlert) {
 
         $scope.errors = {};
         $scope.filters = {  searchFilter: '' };
@@ -76,16 +76,28 @@ angular.module('app')
 
         $scope.delete =  function(camera){
 
-            cameraService.remove(camera._id)
-                .success(function (response, status, headers, config) {
+          SweetAlert.swal({
+              title: "Are you sure?",
+              text: "Your will not be able to recover this camera!",
+              type: "warning",
+              showCancelButton: true,
+              confirmButtonColor: "#DD6B55",confirmButtonText: "Yes, delete it!",
+              cancelButtonText: "No, cancel please!"
+            },
+            function(isConfirm) {
+              if (isConfirm) {
+                cameraService.remove(camera._id)
+                  .success(function (response, status, headers, config) {
                     $scope.tableParams.reload();
-                })
-                .error(function(response, status, headers, config) {
-                    angular.forEach(response.errors, function(error, field) {
-                        form[field].$setValidity('mongoose', false);
-                        $scope.errors[field] = error.type;
+                  })
+                  .error(function (response, status, headers, config) {
+                    angular.forEach(response.errors, function (error, field) {
+                      form[field].$setValidity('mongoose', false);
+                      $scope.errors[field] = error.type;
                     });
-                });
+                  });
+              }
+            });
         }
