+++ public/modules/devices/controllers/device_list.js
@@ -1,7 +1,7 @@
 'use strict';
 
 angular.module('app')
-    .controller('DeviceListCtrl', function ($scope, deviceService, ngTableParams, $modal, $log, psResponsive, $window) {
+    .controller('DeviceListCtrl', function ($scope, deviceService, ngTableParams, $modal, $log, psResponsive, $window, SweetAlert) {
 
         $scope.errors = {};
         $scope.filters = {  searchFilter: '' };
@@ -75,17 +75,29 @@ angular.module('app')
 
         $scope.delete =  function(device){
 
-            deviceService.remove(device._id)
-                .success(function (response, status, headers, config) {
+          SweetAlert.swal({
+              title: "Are you sure?",
+              text: "Your will not be able to recover this device!",
+              type: "warning",
+              showCancelButton: true,
+              confirmButtonColor: "#DD6B55",confirmButtonText: "Yes, delete it!",
+              cancelButtonText: "No, cancel please!"
+            },
+            function(isConfirm) {
+              if (isConfirm) {
+                deviceService.remove(device._id)
+                  .success(function (response, status, headers, config) {
                     $scope.tableParams.reload();
-                })
-                .error(function(response, status, headers, config) {
-                    angular.forEach(response.errors, function(error, field) {
-                        form[field].$setValidity('mongoose', false);
-                        $scope.errors[field] = error.type;
+                  })
+                  .error(function (response, status, headers, config) {
+                    angular.forEach(response.errors, function (error, field) {
+                      form[field].$setValidity('mongoose', false);
+                      $scope.errors[field] = error.type;
                     });
-                });
-        }
+                  });
+              }
+            });
+        };
 
 
         $scope.initDataTable();
