+++ public/modules/triggers/controllers/trigger_list.js
@@ -1,7 +1,7 @@
 'use strict';
 
 angular.module('app')
-    .controller('TriggerListCtrl', function ($scope, triggerService, ngTableParams, $modal, $log, psResponsive,$window) {
+    .controller('TriggerListCtrl', function ($scope, triggerService, ngTableParams, $modal, $log, psResponsive, $window, SweetAlert) {
 
     $scope.errors = {};
     $scope.filters = {  searchFilter: '' };
@@ -76,15 +76,28 @@ angular.module('app')
 
     $scope.delete = function (trigger) {
 
-        triggerService.remove(trigger._id)
-        .success(function (response, status, headers, config) {
-            $scope.tableParams.reload();
-        }).error(function (response, status, headers, config) {
-            angular.forEach(response.errors, function (error, field) {
-                form[field].$setValidity('mongoose', false);
-                $scope.errors[field] = error.type;
-            });
+      SweetAlert.swal({
+          title: "Are you sure?",
+          text: "Your will not be able to recover this trigger!",
+          type: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#DD6B55",confirmButtonText: "Yes, delete it!",
+          cancelButtonText: "No, cancel please!"
+        },
+        function(isConfirm) {
+          if (isConfirm) {
+            triggerService.remove(trigger._id)
+              .success(function (response, status, headers, config) {
+                $scope.tableParams.reload();
+              }).error(function (response, status, headers, config) {
+                angular.forEach(response.errors, function (error, field) {
+                  form[field].$setValidity('mongoose', false);
+                  $scope.errors[field] = error.type;
+                });
+              });
+          }
         });
+
     };
 
     $scope.initDataTable();
