+++ public/modules/users/controllers/user_list.js
@@ -1,7 +1,7 @@
 'use strict';
 
 angular.module('app')
-    .controller('UserListCtrl', function ($scope, userService, ngTableParams, $modal, $log, psResponsive,$window) {
+    .controller('UserListCtrl', function ($scope, userService, ngTableParams, $modal, $log, psResponsive, $window, SweetAlert) {
 
         $scope.errors = {};
         $scope.filters = {  searchFilter: '' };
@@ -74,16 +74,32 @@ angular.module('app')
 
         $scope.delete =  function(device){
 
-            userService.remove(device._id)
-                .success(function (response, status, headers, config) {
+          SweetAlert.swal({
+              title: "Are you sure?",
+              text: "Your will not be able to recover this user!",
+              type: "warning",
+              showCancelButton: true,
+              confirmButtonColor: "#DD6B55",confirmButtonText: "Yes, delete it!",
+              cancelButtonText: "No, cancel please!"
+              },
+            function(isConfirm){
+              if (isConfirm) {
+
+                userService.remove(device._id)
+                  .success(function (response, status, headers, config) {
                     $scope.tableParams.reload();
-                })
-                .error(function(response, status, headers, config) {
+                  })
+                  .error(function(response, status, headers, config) {
                     angular.forEach(response.errors, function(error, field) {
-                        form[field].$setValidity('mongoose', false);
-                        $scope.errors[field] = error.type;
+                      form[field].$setValidity('mongoose', false);
+                      $scope.errors[field] = error.type;
                     });
-                });
+                  });
+
+              }
+            });
+
+
         }
